+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodNoJavadocDefault.java
@@ -70,72 +70,3 @@ public class InputJavadocMethodNoJavadocDefault //comment test
     }
 }
 
-class PackageClass { // ignored
-    public int i1;
-    protected int i2;
-    int i3;
-    private int i4;
-
-    public void foo1() {}
-    protected void foo2() {}
-    void foo3() {}
-    private void foo4() {}
-
-    public class PublicInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    protected class ProtectedInner {
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class PackageInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    private class PrivateInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class IgnoredName { // ignored
-        // ignore by name
-        private int logger;
-        // no warning, 'serialVersionUID' fields do not require Javadoc
-        private static final long serialVersionUID = 0;
-    }
-
-    /**/
-    void methodWithTwoStarComment() {} // ignored
-}
