+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodNoJavadocOnlyPrivateScopePackage.java
@@ -12,8 +12,7 @@ tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
 
-public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
-{
+public class InputJavadocMethodNoJavadocOnlyPrivateScopePackage { // ignored
     public int i1;
     protected int i2;
     int i3;
@@ -24,14 +23,17 @@ public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
     void foo3() {}
     private void foo4() {}
 
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    public void foo5() {}
-    /** @return correct, out of scope */
-    protected void foo6() {}
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    void foo7() {}
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    private void foo8() {}
+    public class PublicInner { // ignored
+        public int i1;
+        protected int i2;
+        int i3;
+        private int i4;
+
+        public void foo1() {}
+        protected void foo2() {}
+        void foo3() {}
+        private void foo4() {}
+    }
 
     protected class ProtectedInner { // ignored
         public int i1;
@@ -68,5 +70,14 @@ public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
         void foo3() {}
         private void foo4() {}
     }
-}
 
+    class IgnoredName { // ignored
+        // ignore by name
+        private int logger;
+        // no warning, 'serialVersionUID' fields do not require Javadoc
+        private static final long serialVersionUID = 0;
+    }
+
+    /**/
+    void methodWithTwoStarComment() {}  // ignored
+}
