+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodNoJavadocProtectedScope.java
@@ -69,73 +69,3 @@ public class InputJavadocMethodNoJavadocProtectedScope //comment test
         private void foo4() {}
     }
 }
-
-class InputJavadocMethodNoJavadocProtectedScopePackage { // ignored
-    public int i1;
-    protected int i2;
-    int i3;
-    private int i4;
-
-    public void foo1() {}
-    protected void foo2() {}
-    void foo3() {}
-    private void foo4() {}
-
-    public class PublicInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    protected class ProtectedInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class PackageInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    private class PrivateInner { // ignored
-        public int i1;
-        protected int i2;
-        int i3;
-        private int i4;
-
-        public void foo1() {}
-        protected void foo2() {}
-        void foo3() {}
-        private void foo4() {}
-    }
-
-    class IgnoredName { // ignored
-        // ignore by name
-        private int logger;
-        // no warning, 'serialVersionUID' fields do not require Javadoc
-        private static final long serialVersionUID = 0;
-    }
-
-    /**/
-    void methodWithTwoStarComment() {} // ignored
-}
