+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodPackageClass.java
@@ -2,7 +2,7 @@
 JavadocMethod
 allowedAnnotations = (default)Override
 validateThrows = (default)false
-accessModifiers = private, package, public
+accessModifiers = (default)public, protected, package, private
 allowMissingParamTags = (default)false
 allowMissingReturnTag = (default)false
 tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
@@ -12,8 +12,7 @@ tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
 
-public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
-{
+public class InputJavadocMethodPackageClass { // ignored
     public int i1;
     protected int i2;
     int i3;
@@ -24,16 +23,19 @@ public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
     void foo3() {}
     private void foo4() {}
 
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    public void foo5() {}
-    /** @return correct, out of scope */
-    protected void foo6() {}
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    void foo7() {}
-    /** @return wrong, in scope */ // violation 'Unused Javadoc tag.'
-    private void foo8() {}
+    public class PublicInner { // ignored
+        public int i1;
+        protected int i2;
+        int i3;
+        private int i4;
 
-    protected class ProtectedInner { // ignored
+        public void foo1() {}
+        protected void foo2() {}
+        void foo3() {}
+        private void foo4() {}
+    }
+
+    protected class ProtectedInner {
         public int i1;
         protected int i2;
         int i3;
@@ -68,5 +70,14 @@ public class InputJavadocMethodNoJavadocOnlyPrivateScope //comment test
         void foo3() {}
         private void foo4() {}
     }
-}
 
+    class IgnoredName { // ignored
+        // ignore by name
+        private int logger;
+        // no warning, 'serialVersionUID' fields do not require Javadoc
+        private static final long serialVersionUID = 0;
+    }
+
+    /**/
+    void methodWithTwoStarComment() {} // ignored
+}
