+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodPackageClass2.java
@@ -10,10 +10,10 @@ tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
 
 */
 
+
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
 
-public class InputJavadocMethodSurroundingAccessModifier //comment test
-{
+public class InputJavadocMethodPackageClass2 {
     public int i1;
     protected int i2;
     int i3;
@@ -31,28 +31,23 @@ public class InputJavadocMethodSurroundingAccessModifier //comment test
     private void foo4() {
     }
 
-    /**
-     * @return
-     */
-    public void foo5() {
-    }
+    public class PublicInner {
+        public int i1;
+        protected int i2;
+        int i3;
+        private int i4;
 
-    /**
-     * @return
-     */
-    protected void foo6() {
-    }
+        public void foo1() {
+        }
 
-    /**
-     * @return
-     */
-    void foo7() {
-    }
+        protected void foo2() {
+        }
 
-    /**
-     * @return
-     */
-    private void foo8() {
+        void foo3() {
+        }
+
+        private void foo4() {
+        }
     }
 
     protected class ProtectedInner {
@@ -111,5 +106,15 @@ public class InputJavadocMethodSurroundingAccessModifier //comment test
         private void foo4() {
         }
     }
-}
 
+    class IgnoredName {
+        // ignore by name
+        private int logger;
+        // no warning, 'serialVersionUID' fields do not require Javadoc
+        private static final long serialVersionUID = 0;
+    }
+
+    /**/
+    void methodWithTwoStarComment() {
+    }
+}
