+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodProtectedScopeJavadocOne.java
@@ -1,0 +1,56 @@
+/*
+JavadocMethod
+allowedAnnotations = (default)Override
+validateThrows = (default)false
+accessModifiers = public, protected
+allowMissingParamTags = (default)false
+allowMissingReturnTag = (default)false
+tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
+
+public class InputJavadocMethodProtectedScopeJavadocOne // ignore - need javadoc
+{
+    private interface InnerInterface // ignore - when not relaxed about Javadoc
+    {
+        String CONST = "InnerInterface"; // ignore - w.n.r.a.j
+        void method(); // ignore - when not relaxed about Javadoc
+
+        class InnerInnerClass // ignore - when not relaxed about Javadoc
+        {
+            private int mData; // ignore - when not relaxed about Javadoc
+
+            private InnerInnerClass()
+            {
+                final Runnable r = new Runnable() {
+                    public void run() {};
+                };
+            }
+
+            void method2() // ignore - when not relaxed about Javadoc
+            {
+                final Runnable r = new Runnable() {
+                    public void run() {};
+                };
+            }
+        }
+    }
+
+    private class InnerClass // ignore
+    {
+        private int mDiff; // ignore - when not relaxed about Javadoc
+
+        void method() // ignore - when not relaxed about Javadoc
+        {
+        }
+    }
+
+    private int mSize; // ignore - when not relaxed about Javadoc
+    int mLen; // ignore - when not relaxed about Javadoc
+    protected int mDeer; // ignore
+    public int aFreddo; // ignore
+
+}
