+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodProtectedScopeJavadocTwo.java
@@ -12,64 +12,24 @@ tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
 
-public class InputJavadocMethodProtectedScopeJavadoc // ignore - need javadoc
-{
-    private interface InnerInterface // ignore - when not relaxed about Javadoc
-    {
-        String CONST = "InnerInterface"; // ignore - w.n.r.a.j
-        void method(); // ignore - when not relaxed about Javadoc
-
-        class InnerInnerClass // ignore - when not relaxed about Javadoc
-        {
-            private int mData; // ignore - when not relaxed about Javadoc
-
-            private InnerInnerClass()
-            {
-                final Runnable r = new Runnable() {
-                    public void run() {};
-                };
-            }
-
-            void method2() // ignore - when not relaxed about Javadoc
-            {
-                final Runnable r = new Runnable() {
-                    public void run() {};
-                };
-            }
-        }
-    }
-
-    private class InnerClass // ignore
-    {
-        private int mDiff; // ignore - when not relaxed about Javadoc
-
-        void method() // ignore - when not relaxed about Javadoc
-        {
-        }
-    }
-
-    private int mSize; // ignore - when not relaxed about Javadoc
-    int mLen; // ignore - when not relaxed about Javadoc
-    protected int mDeer; // ignore
-    public int aFreddo; // ignore
-
-    // ignore - need Javadoc
-    private InputJavadocMethodProtectedScopeJavadoc(int aA)
+public class InputJavadocMethodProtectedScopeJavadocTwo {
+        // ignore - need Javadoc
+    private InputJavadocMethodProtectedScopeJavadocTwo(int aA)
     {
     }
 
     // ignore - need Javadoc when not relaxed
-    InputJavadocMethodProtectedScopeJavadoc(String aA)
+    InputJavadocMethodProtectedScopeJavadocTwo(String aA)
     {
     }
 
     // ignore - always need javadoc
-    protected InputJavadocMethodProtectedScopeJavadoc(Object aA)
+    protected InputJavadocMethodProtectedScopeJavadocTwo(Object aA)
     {
     }
 
     // ignore - always need javadoc
-    public InputJavadocMethodProtectedScopeJavadoc(Class<Object> aA)
+    public InputJavadocMethodProtectedScopeJavadocTwo(Class<Object> aA)
     {
     }
