+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodPublicOnly1Two.java
@@ -1,0 +1,57 @@
+/*
+JavadocMethod
+allowedAnnotations = (default)Override
+validateThrows = (default)false
+accessModifiers =
+allowMissingParamTags = (default)false
+allowMissingReturnTag = (default)false
+tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
+
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
+
+public class InputJavadocMethodPublicOnly1Two {
+    /** {@inheritDoc} */
+    public String toString()
+    {
+        return super.toString();
+    }
+
+    @Deprecated @Override
+    public int hashCode()
+    {
+        return super.hashCode();
+    }
+
+    public Thread anonymousClassInMethod() {
+        return new Thread() {
+            @Override
+            public void run() {
+                privateMethod(null, null);
+            }
+
+            /**
+             * Javadoc
+             */
+            private String privateMethod(String a, String b) {
+                return null;
+            }
+        };
+    }
+
+    private final Thread anonymousClassInField = new Thread() {
+        @Override
+        public void run() {
+            publicMethod(null, null);
+        }
+
+        /**
+         * Javadoc
+         */
+        public String publicMethod(String a, String b) {
+            return null;
+        }
+    };
+}
