+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodPublicOnlyOne.java
@@ -2,7 +2,7 @@
 JavadocMethod
 allowedAnnotations = (default)Override
 validateThrows = (default)false
-accessModifiers =
+accessModifiers = (default)public, protected, package, private
 allowMissingParamTags = (default)false
 allowMissingReturnTag = (default)false
 tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
@@ -12,7 +12,7 @@ tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
 
 package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
 
-public class InputJavadocMethodPublicOnly1 // ignore - need javadoc
+public class InputJavadocMethodPublicOnlyOne // ignore - need javadoc
 {
     private interface InnerInterface // ignore - when not relaxed about Javadoc
     {
@@ -54,102 +54,43 @@ public class InputJavadocMethodPublicOnly1 // ignore - need javadoc
     public int aFreddo; // ignore
 
     // ignore - need Javadoc
-    private InputJavadocMethodPublicOnly1(int aA)
+    private InputJavadocMethodPublicOnlyOne(int aA)
     {
     }
 
     // ignore - need Javadoc when not relaxed
-    InputJavadocMethodPublicOnly1(String aA)
+   InputJavadocMethodPublicOnlyOne(String aA)
     {
     }
 
     // ignore - always need javadoc
-    protected InputJavadocMethodPublicOnly1(Object aA)
+    protected InputJavadocMethodPublicOnlyOne(Object aA)
     {
     }
 
     // ignore - always need javadoc
-    public InputJavadocMethodPublicOnly1(Class<Object> aA)
+   public InputJavadocMethodPublicOnlyOne(Class<Object> aA)
     {
     }
 
     /** Here should be an error, In scope */
-    private void method(int aA)
+    private void method(int aA) // violation 'Expected @param tag for 'aA'.'
     {
     }
 
     /** Here should be an error, In scope */
-    void method(Long aA)
+    void method(Long aA) // violation 'Expected @param tag for 'aA'.'
     {
     }
 
     /** Here should be an error, In scope */
-    protected void method(Class<Object> aA)
+    protected void method(Class<Object> aA) // violation 'Expected @param tag for 'aA'.'
     {
     }
 
     /** Here should be an error, In scope */
-    public void method(StringBuffer aA)
+    public void method(StringBuffer aA) // violation 'Expected @param tag for 'aA'.'
     {
     }
 
-
-    /**
-       A param tag should not be required here when relaxed about Javadoc.
-       Writing a little documentation should not be worse than not
-       writing any documentation at all.
-     */
-    private void method(String aA)
-    {
-    }
-
-    /**
-       This inner class has no author tag, which is OK.
-     */
-    public class InnerWithoutAuthor
-    {
-
-    }
-
-    /** {@inheritDoc} */
-    public String toString()
-    {
-        return super.toString();
-    }
-
-    @Deprecated @Override
-    public int hashCode()
-    {
-        return super.hashCode();
-    }
-
-    public Thread anonymousClassInMethod() {
-        return new Thread() {
-            @Override
-            public void run() {
-                privateMethod(null, null);
-            }
-
-            /**
-             * Javadoc
-             */
-            private String privateMethod(String a, String b) {
-                return null;
-            }
-        };
-    }
-
-    private final Thread anonymousClassInField = new Thread() {
-        @Override
-        public void run() {
-            publicMethod(null, null);
-        }
-
-        /**
-         * Javadoc
-         */
-        public String publicMethod(String a, String b) {
-            return null;
-        }
-    };
 }
