+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadocmethod/InputJavadocMethodTags1Three.java
@@ -1,0 +1,117 @@
+/*
+JavadocMethod
+allowedAnnotations = (default)Override
+validateThrows = true
+accessModifiers = (default)public, protected, package, private
+allowMissingParamTags = (default)false
+allowMissingReturnTag = (default)false
+tokens = (default)METHOD_DEF, CTOR_DEF, ANNOTATION_FIELD_DEF, COMPACT_CTOR_DEF
+
+*/
+
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadocmethod;
+
+import java.io.IOException;
+
+public class InputJavadocMethodTags1Three {
+    /**
+     * Bug 579190, "expected return tag when one is there".
+     *
+     * Linebreaks after return tag should be legal.
+     *
+     * @return
+     *   the bug that states that linebreak should be legal
+     */
+    int method20()
+    {
+        return 579190;
+    }
+
+    /**
+     * Bug XXXX, "two tags for the same exception"
+     *
+     * @exception java.io.IOException for some reasons
+     * @exception IOException for another reason
+     */
+    void method21()
+       throws IOException
+    {
+    }
+
+    /**
+     * RFE 540383, "Unused throws tag for exception subclass"
+     *
+     * @exception IOException for some reasons
+     * @exception java.io.FileNotFoundException for another reasons
+     */
+    void method22()
+       throws IOException
+    {
+    }
+
+    /**
+     * @exception WrongException exception w/o class info but matched by name
+     */
+    void method23() throws WrongException
+    {
+    }
+
+    /**
+     * Bug 803577, "allowThrowsTagsForSubclasses/allowMissingThrowsTag interfere"
+     *
+     * no exception tag for IOException, but here is a tag for its subclass.
+     * @exception java.io.FileNotFoundException for another reasons
+     */
+    void method24() throws IOException // violation 'Expected @throws tag for 'IOException'.'
+    {
+    }
+
+    /**
+     * Bug 841942, "ArrayIndexOutOfBounds in JavadocStyle".
+     * @param aParam there is no such param in the method.
+     * The problem should be reported with correct line number.
+     */
+    // violation 3 lines above 'Un.* @param .* 'aParam'.'
+    void method25()
+    {
+    }
+
+    /** {@inheritDoc} */
+    int method26()
+    { return 0;
+    }
+
+    /**
+     * {@inheritDoc}
+     * @return something very important.
+     */
+    int method27(int aParam)
+    { return 0;
+    }
+
+    /**
+     * @return something very important.
+     * {@inheritDoc}
+     */
+    int method28(int aParam)
+    { return 0;
+    }
+
+    /**
+     * {@inheritDoc}
+     *
+     * @return 1
+     */
+    public int foo(Object _arg) {
+
+        return 1;
+    }
+
+    /**
+     * misplaced @param aParam
+     * misplaced @return something very important.
+     */
+    int method29(int aParam) // 2 violations
+    { return 0;
+    }
+}
